+++ sdk/sui/timelock_configurer.go
@@ -74,3 +74,13 @@ func serializeTimelockUpdateMinDelay(timelockAddress string, newMinDelay uint64)
 		ser.U64(newMinDelay)
 	})
 }
+
+// GrantRole grants a timelock role to an address.
+func (c *TimelockConfigurer) GrantRole(
+	ctx context.Context,
+	timelockAddress string,
+	role sdk.TimelockRole,
+	targetAddress string,
+) (types.TransactionResult, error) {
+	panic("not implemented")
+}